// src/core/enums/error-codes.enum.ts
/**
 * File: error-codes.enum.ts
 * Project: nest-microservice-boilerplate
 * Version:1.0.0
 * Created Date: Saturday, February 1st 2020, 1:24:51 pm
 * Author: Georgian Stan (georgian.stan8@gmail.com)
 * -----
 * Last Modified: Saturday, 1st February 2020 1:51:50 pm
 * Modified By: Georgian Stan (georgian.stan8@gmail.com>)
 * ------------------------------------
 * Javascript will save your soul!
 */

export enum ErrCodes {
  UNKNOWN = 1000,
  DATABASE_ERROR = 1001,
  BAD_PARAMETERS = 1100,
}